lax_numpy: move array and asarray to their own submodule
This leads to a better dependency graph and avoids local imports.
