+++ jax/_src/numpy/scalar_types.py
@@ -27,6 +27,7 @@ import numpy as np
 from jax._src.typing import Array
 from jax._src import core
 from jax._src import dtypes
+from jax._src.numpy.array import asarray
 
 
 # Some objects below rewrite their __module__ attribute to this name.
@@ -46,7 +47,6 @@ class _ScalarMeta(type):
     return not (self == other)
 
   def __call__(self, x: Any) -> Array:
-    from jax._src.numpy.lax_numpy import asarray
     return asarray(x, dtype=self.dtype)
 
   def __instancecheck__(self, instance: Any) -> bool:
