+++ jax/numpy/__init__.py
@@ -24,6 +24,11 @@ from jax._src.dtypes import (
     isdtype as isdtype,
 )
 
+from jax._src.numpy.array import (
+    array as array,
+    asarray as asarray,
+)
+
 from jax._src.numpy.lax_numpy import (
     ComplexWarning as ComplexWarning,
     allclose as allclose,
@@ -36,12 +41,10 @@ from jax._src.numpy.lax_numpy import (
     argmin as argmin,
     argwhere as argwhere,
     around as around,
-    array as array,
     array_equal as array_equal,
     array_equiv as array_equiv,
     array_split as array_split,
     astype as astype,
-    asarray as asarray,
     atleast_1d as atleast_1d,
     atleast_2d as atleast_2d,
     atleast_3d as atleast_3d,
